[PC-10266] OfferForm: extract accessibility checkboxes
