+++ src/components/pages/Offers/Offer/OfferDetails/OfferForm/AccessibilityCheckboxList/index.ts
@@ -1,0 +1,1 @@
+export { default } from './AccessibilityCheckboxList'
